feat: rename TopBarV2 components to TopBar
